[refactor] #16 기타 api들 하나의 이미지 반환 api로 통합
