--- src/main/java/com/burger/burgerking/story/api/BrandController.java
@@ -1,35 +1,0 @@
-package com.burger.burgerking.story.api;
-
-import com.burger.burgerking.global.common.ApiResponse;
-import com.burger.burgerking.story.application.BrandService;
-import com.burger.burgerking.story.dto.request.BrandRequest;
-import com.burger.burgerking.story.dto.response.BrandResponse;
-import com.burger.burgerking.story.dto.response.FileMetaDataResponse;
-import com.burger.burgerking.story.dto.response.wrapper.BrandAndFileMetaDataResponse;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@Tag(name = "Brand", description = "Brand 섹션")
-@RestController
-@RequestMapping("/api/v1/story/brand")
-@RequiredArgsConstructor
-public class BrandController {
-
-    private final BrandService brandService;
-
-    @Operation(summary = "Brand 정보 제공", description = "Brand에 대한 title, description, imageUrl 제공")
-    @PostMapping
-    public ResponseEntity<ApiResponse<List<FileMetaDataResponse>>> brand() {
-
-        return ResponseEntity.ok(ApiResponse.success(brandService.getInfo()));
-    }
-}
