--- src/main/java/com/burger/burgerking/story/api/WhyBurgerkingController.java
@@ -1,35 +1,0 @@
-package com.burger.burgerking.story.api;
-
-import com.burger.burgerking.global.common.ApiResponse;
-import com.burger.burgerking.story.application.WhyBurgerKingService;
-import com.burger.burgerking.story.dto.request.WhyBurgerKingRequest;
-import com.burger.burgerking.story.dto.response.BrandResponse;
-import com.burger.burgerking.story.dto.response.FileMetaDataResponse;
-import com.burger.burgerking.story.dto.response.WhyResponse;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@Tag(name = "Why Burgerking", description = "Why Burgerking 섹션")
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/story/why")
-public class WhyBurgerkingController {
-
-    private final WhyBurgerKingService whyBurgerKingService;
-
-    @Operation(summary = "Why Burgerking 정보 제공", description = "Why Burgerking에 대한 title, subtitle, description, imageUrl 제공")
-    @PostMapping
-    public ResponseEntity<ApiResponse<List<FileMetaDataResponse>>> why(@RequestBody WhyBurgerKingRequest request) {
-
-        return ResponseEntity.ok(ApiResponse.success(whyBurgerKingService.getInfo(request)));
-    }
-}
