--- src/main/java/com/burger/burgerking/story/dao/WhyBurgerKingRepository.java
@@ -1,8 +1,0 @@
-package com.burger.burgerking.story.dao;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class WhyBurgerKingRepository {
-
-}
