+++ src/main/java/com/burger/burgerking/story/dto/request/EsgRequest.java
@@ -1,0 +1,8 @@
+package com.burger.burgerking.story.dto.request;
+
+import com.burger.burgerking.story.enums.FileType;
+
+public record EsgRequest(
+        FileType type
+) {
+}
