+++ src/main/java/com/burger/burgerking/story/dto/request/ImageRequest.java
@@ -1,0 +1,8 @@
+package com.burger.burgerking.story.dto.request;
+
+import com.burger.burgerking.story.enums.FileType;
+
+public record ImageRequest(
+        FileType type
+) {
+}
